split example code
