+++ example/example.labelfixed.css
@@ -1,0 +1,16 @@
+body.labelFixed .control {
+  flex-direction: row;
+}
+body.labelFixed .control > label {
+  min-width: 10em;
+  text-align: end;
+  margin-right: 1em;
+}
+body.labelFixed .control > .input {
+  flex: 1;
+}
+/* Workaround for https://bugs.chromium.org/p/chromium/issues/detail?id=375693 */
+/*body.labelFixed .control > fieldset {
+  display: block;
+  width: 100%;
+}*/
