+++ example/index.ts
@@ -1,4 +1,6 @@
-import './example';
+import {createExampleSelection} from './example';
+import {createThemeSelection} from './theme.switcher';
+import {createStyleSelection} from './style.switcher';
 import './templates/arrays';
 import './templates/day2';
 import './templates/day4';
@@ -15,3 +17,9 @@ import './templates/categorization';
 import './templates/masterdetail';
 import './templates/resolve';
 import './templates/uischema-registry';
+
+window.onload = (ev) => {
+  const selectExampleElement = createExampleSelection();
+  createThemeSelection();
+  createStyleSelection(selectExampleElement);
+};
