+++ example_plain/index.html
@@ -19,6 +19,8 @@
 
   <!-- Plain CSS -->
   <link rel="stylesheet" href="assets/example.css">
+  <link rel="stylesheet" href="assets/example.dark.css">
+  <link rel="stylesheet" href="assets/example.labelfixed.css">
 
   <!-- Bootstrap JS, if we need it, it must be loaded before materialize JS, otherwise the select elements seem to break -->
   <!--<script src="assets/bootstrap.js"></script>-->
